dashboard/data/requirements.txt
flask
requests
Flask-Session
cryptography
werkzeug==0.16.0
websocket-client